Added alpha-beta prunning and position analyzer

+++ src/main/java/core_v2/Evaluators/Evaluator.java
@@ -8,5 +8,5 @@ import core_v2.Chessboards.IBoard;
  */
 public interface Evaluator {
 
-    double evaluate(Chessboard chessboard, int depth);
+    int evaluate(Chessboard chessboard, int depth);
 }
